- DiffEntry constructor was building a String on every invocation
  just for Preconditions.checkArgument, reported as 30% of total
  time by the profiler.
- CheckMergeScenarioOp unnecessarily called ResolveObjectType.class
- Index.stage() called ConflictsDatabase.removeConflict() even if there
  were no conflicts.
- ReportMergeScenarioOp called RevObjectParse for the same arguments
  inside a loop
- ReportMergeScenarioOp called ResolveObjectType unnecessarily
- MergeOp could build a very large exception message, one line per
  conflict. Limited to 50 now.

+++ src/core/src/main/java/org/locationtech/geogig/repository/Index.java
@@ -175,7 +175,8 @@ public class Index implements StagingArea {
         Map<String, RevTreeBuilder> parentTress = Maps.newHashMap();
         Map<String, ObjectId> parentMetadataIds = Maps.newHashMap();
         Set<String> removedTrees = Sets.newHashSet();
-        ConflictsDatabase conflictsDb = conflictsDatabase();
+        final ConflictsDatabase conflictsDb = conflictsDatabase();
+        final boolean hasConflicts = conflictsDb.hasConflicts(null);
         while (unstaged.hasNext()) {
             final DiffEntry diff = unstaged.next();
             final String fullPath = diff.oldPath() == null ? diff.newPath() : diff.oldPath();
@@ -219,8 +220,9 @@ public class Index implements StagingArea {
                 Node node = newObject.getNode();
                 parentTree.put(node);
             }
-
-            conflictsDb.removeConflict(null, fullPath);
+            if (hasConflicts) {
+                conflictsDb.removeConflict(null, fullPath);
+            }
         }
 
         ObjectId newRootTree = currentIndexHead.getId();
@@ -273,8 +275,8 @@ public class Index implements StagingArea {
 
                 parentBuilder = new RevTreeBuilder(context.objectDatabase(),
                         context.command(FindOrCreateSubtree.class)
-                        .setParent(Suppliers.ofInstance(Optional.of(getTree())))
-                        .setChildPath(parentPath).call());
+                                .setParent(Suppliers.ofInstance(Optional.of(getTree())))
+                                .setChildPath(parentPath).call());
             }
             parentTress.put(parentPath, parentBuilder);
             if (parentMetadataId != null) {
